+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Change Log
 
 This document records the main changes to the sdss_access code.
 
-3.0.4 (unreleased)
+3.0.4 (03-08-2024)
 ------------------
 - Fix issue `52` - rsync failure when remote file is compressed compared to template
 - Issue `48` - Add support for adding temporary paths for use in local sdss_access